+++ package.json
@@ -10,19 +10,17 @@
     "prepare": "husky install",
     "lint": "eslint --fix --ext .ts,.tsx ./src"
   },
-  
   "lint-staged": {
     "**/*.{js,jsx,ts,tsx}": [
       "npm run lint"
     ]
-},
-
-
+  },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
     "firebase": "^9.22.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.1"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
