--- src/components/Aunthentcation /SignUpFacebook.tsx
@@ -1,9 +1,0 @@
-// import React from 'react'
-
-const SignUpFacebook = () => {
-  return (
-    <div>SignUpFacebook</div>
-  )
-}
-
-export default SignUpFacebook
