+++ src/components/Aunthentcation /SigninFacebook.tsx
@@ -1,0 +1,9 @@
+// import React from 'react'
+
+const SigninFacebook = () => {
+  return (
+    <div>SigninFacebook</div>
+  )
+}
+
+export default SigninFacebook
