+++ src/components/Aunthentcation /SigninGoogle.tsx
@@ -1,0 +1,9 @@
+// import React from 'react'
+
+const SigninGoogle = () => {
+  return (
+    <div>SigninGoogle</div>
+  )
+}
+
+export default SigninGoogle
