+++ src/components/Counter.tsx
@@ -1,9 +1,18 @@
-// import React from 'react'
+import { useSelector, useDispatch } from 'react-redux';
+import { increment, decrement } from '../store/counterSlice';
 
 const Counter = () => {
-  return (
-    <div className='text-3xl font-bold text-blue-600' >Welcome to Diary App </div>
-  )
-}
+  const count = useSelector((state) => state.counter.value);
+  const dispatch = useDispatch();
 
-export default Counter
+  return (
+    <div>
+       <div className='text-3xl font-bold text-blue-600' >Welcome to Diary App </div>
+      <div>Count: {count}</div>
+      <button onClick={() => dispatch(increment())}>Increment</button>
+      <button onClick={() => dispatch(decrement())}>Decrement</button>
+    </div>
+  );
+};
+
+export default Counter;
